requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Data & Backtesting
pandas==2.1.3
numpy==1.26.2
vectorbt==0.25.4
yfinance==0.2.32
ta-lib==0.4.28

# Trading & Exchange Integration
ccxt==4.0.30
stable-baselines3==2.2.1

# Machine Learning & AI
torch==2.1.1
transformers==4.35.2
scikit-learn==1.3.2

# Database & Caching
redis==5.0.1
sqlalchemy-utils==0.41.1

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0
requests==2.31.0
python-telegram-bot==20.3

# Monitoring & Logging
python-json-logger==2.0.7
mlflow==2.10.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0

# Async Support
aiohttp==3.9.1
asyncio==3.4.3

# Terminal UI
textual==0.45.1
click==8.1.7
rich==13.7.0
